requirements.txt
fastapi
uvicorn
stellar-sdk
supabase
python-dotenv
httpx